refactor(stripe): migrate Stripe service methods to a dedicated StripeService class and remove third-party service file
